+++ examples/shared_vpc/main.tf
@@ -48,7 +48,6 @@ module "host-project" {
   org_id                         = var.organization_id
   folder_id                      = var.folder_id
   billing_account                = var.billing_account
-  skip_gcloud_download           = true
   enable_shared_vpc_host_project = true
 }
 
@@ -57,7 +56,7 @@ module "host-project" {
  *****************************************/
 module "vpc" {
   source  = "terraform-google-modules/network/google"
-  version = "~> 2.1.0"
+  version = "~> 2.5.0"
 
   project_id                             = module.host-project.project_id
   network_name                           = var.network_name
@@ -103,15 +102,14 @@ module "vpc" {
   Service Project Creation
  *****************************************/
 module "service-project" {
-  source = "../../modules/shared_vpc"
+  source = "../../modules/svpc_service_project"
 
   name              = var.service_project_name
   random_project_id = "false"
 
-  org_id             = var.organization_id
-  folder_id          = var.folder_id
-  billing_account    = var.billing_account
-  shared_vpc_enabled = true
+  org_id          = var.organization_id
+  folder_id       = var.folder_id
+  billing_account = var.billing_account
 
   shared_vpc         = module.host-project.project_id
   shared_vpc_subnets = module.vpc.subnets_self_links
@@ -124,22 +122,20 @@ module "service-project" {
   ]
 
   disable_services_on_destroy = "false"
-  skip_gcloud_download        = "true"
 }
 
 /******************************************
   Second Service Project Creation
  *****************************************/
 module "service-project-b" {
-  source = "../../modules/shared_vpc"
+  source = "../../modules/svpc_service_project"
 
   name              = "b-${var.service_project_name}"
   random_project_id = "false"
 
-  org_id             = var.organization_id
-  folder_id          = var.folder_id
-  billing_account    = var.billing_account
-  shared_vpc_enabled = true
+  org_id          = var.organization_id
+  folder_id       = var.folder_id
+  billing_account = var.billing_account
 
   shared_vpc = module.host-project.project_id
 
@@ -149,8 +145,15 @@ module "service-project-b" {
     "dataproc.googleapis.com",
   ]
 
+  activate_api_identities = [{
+    api = "healthcare.googleapis.com"
+    roles = [
+      "roles/healthcare.serviceAgent",
+      "roles/bigquery.jobUser",
+    ]
+  }]
+
   disable_services_on_destroy = "false"
-  skip_gcloud_download        = "true"
 }
 
 /******************************************