+++ modules/core_project_factory/versions.tf
@@ -15,12 +15,24 @@
  */
 
 terraform {
-  required_version = ">=0.12.6, <0.14"
+  required_version = ">= 0.13"
 
   required_providers {
-    google      = ">= 3.8, < 4.0"
-    google-beta = ">= 3.8, < 4.0"
-    null        = "~> 2.1"
-    random      = "~> 2.2"
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.50, < 4.0"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = ">= 3.50, < 4.0"
+    }
+    null = {
+      source  = "hashicorp/null"
+      version = "~> 2.1"
+    }
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 2.2"
+    }
   }
 }