+++ modules/project_services/main.tf
@@ -15,7 +15,7 @@
  */
 
 locals {
-  services = toset(concat(var.activate_apis, [for i in var.activate_api_identities : i.api]))
+  services = var.enable_apis ? toset(concat(var.activate_apis, [for i in var.activate_api_identities : i.api])) : toset([])
   service_identities = flatten([
     for i in var.activate_api_identities : [
       for r in i.roles :