+++ modules/project_services/versions.tf
@@ -15,9 +15,21 @@
  */
 
 terraform {
-  required_version = ">=0.12.6, <0.14"
-
+  required_version = ">= 0.13"
   required_providers {
-    google = ">= 2.1, < 4.0"
+    google = {
+      source  = "hashicorp/google"
+      version = ">= 3.43, <4.0"
+    }
+    google-beta = {
+      source  = "hashicorp/google-beta"
+      version = ">= 3.43, <4.0"
+    }
+  }
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v10.2.1"
+  }
+  provider_meta "google-beta" {
+    module_name = "blueprints/terraform/terraform-google-project-factory:project_services/v10.2.1"
   }
 }