+++ modules/shared_vpc_access/variables.tf
@@ -19,9 +19,26 @@ variable "host_project_id" {
   type        = string
 }
 
+variable "enable_shared_vpc_service_project" {
+  description = "Flag set if SVPC enabled"
+  type        = bool
+}
+
 variable "service_project_id" {
   description = "The ID of the service project"
   type        = string
+}
+
+variable "service_project_number" {
+  description = "Project number of the service project. Will be used if `lookup_service_project_number` is false."
+  type        = string
+  default     = null
+}
+
+variable "lookup_project_numbers" {
+  description = "Whether to look up the project numbers from data sources. If false, `service_project_number` will be used instead."
+  type        = bool
+  default     = true
 }
 
 variable "shared_vpc_subnets" {
@@ -34,4 +51,10 @@ variable "active_apis" {
   description = "The list of active apis on the service project. If api is not active this module will not try to activate it"
   type        = list(string)
   default     = []
+}
+
+variable "grant_services_security_admin_role" {
+  description = "Whether or not to grant Kubernetes Engine Service Agent the Security Admin role on the host project so it can manage firewall rules"
+  type        = bool
+  default     = false
 }