+++ versions.tf
@@ -15,5 +15,11 @@
  */
 
 terraform {
-  required_version = ">=0.12.6, <0.14"
+  required_version = ">=0.13.0"
+  provider_meta "google" {
+    module_name = "blueprints/terraform/terraform-google-project-factory/v10.2.1"
+  }
+  provider_meta "google-beta" {
+    module_name = "blueprints/terraform/terraform-google-project-factory/v10.2.1"
+  }
 }